Restructure + disable check on main (#10)
* Seperated ScenarioInfo, TraceInfo, and ScnearioGraph into seperate file according to architecture

* Don't run automated testing on main

* fixing imports
